Revert "Add ppaquette Et al. supervised model Encoder (#8)"
This reverts commit 099f39f49d467bc2044a15c2b67f1355925e4e2a.
