+++ rl_projects/rl_ppaquette_project/dataset_creation.py
@@ -7,34 +7,33 @@ from queue import Queue
 from tqdm import tqdm
 
 
-from utilities.dataset_creation_functions import (
-    get_request_id,
-    get_policy_data,
-    format_datapoint,
-    generate_datapoint,
-)
+from utilities.dataset_creation_functions import (get_request_id, 
+                                                 get_policy_data,
+                                                 format_datapoint,
+                                                 generate_datapoint)
 
 from utilities.data_extraction_functions import decompress_game
 from utilities.utility_functions import compress_dict, decompress_dict
 
-from settings import (
-    DATASET_CREATION_NUM_WORKERS,
-    DATASET_CREATION_BUFFER_SIZE,
-    MODEL_DATA_PATHS,
-    MODEL_DATA_DIR,
-    DATA_FEATURES,
-    H_PARAMS,
-)
+from settings import (MODEL_DATA_PATHS, MODEL_DATA_DIR,
+EXTRACTED_DATA_DIR, VALIDATION_SET_SPLIT,
+N_LOCATIONS, N_SUPPLY_CENTERS,
+N_LOCATION_FEATURES, N_ORDERS_FEATURES,
+N_POWERS, N_SEASONS,
+N_UNIT_TYPES, N_NODES,
+TOKENS_PER_ORDER, MAX_LENGTH_ORDER_PREV_PHASES,
+MAX_CANDIDATES, N_PREV_ORDERS, N_PREV_ORDERS_HISTORY,
+DATASET_CREATION_NUM_WORKERS, DATASET_CREATION_BUFFER_SIZE)
 
 LOGGER = logging.getLogger(__name__)
 
 
 def process_sample(compressed_game_string):
-    """Handles the datapoint processing"""
+    """ Handles the datapoint processing"""
     try:
         saved_game = decompress_game(compressed_game_string)
         is_validation_sample = bool(saved_game["id"] in VALID_IDS)
-
+        
         game_results = generate_datapoint(saved_game, is_validation_sample)
         results = []
         for phase_idx in range(len(saved_game["phases"])):
@@ -47,15 +46,16 @@ def process_sample(compressed_game_string):
         return (saved_game["id"], results)
     except Exception as exc:
         raise exc
-
+        
 
 # train/val datapoint counts
-dataset_index = {"size_train_dataset": 0, "size_valid_dataset": 0}
+dataset_index = {
+    'size_train_dataset' : 0,
+    'size_valid_dataset': 0
+}
 
 # Splitting games into training and validation
-phase_count_dataset = pickle.load(
-    open(MODEL_DATA_PATHS["PHASES_COUNT_DATASET_PATH"], "rb")
-)
+phase_count_dataset = pickle.load(open(MODEL_DATA_PATHS["PHASES_COUNT_DATASET_PATH"], 'rb'))      
 game_ids = list(phase_count_dataset.keys())
 
 n_valid = int(VALIDATION_SET_SPLIT * len(game_ids))
@@ -77,46 +77,40 @@ pending_phases = sum(phase_count_dataset.get(game_id, 0) for game_id in game_ids
 progress_bar = tqdm(total=pending_phases)
 
 # main loop
-with concurrent.futures.ProcessPoolExecutor(
-    max_workers=DATASET_CREATION_NUM_WORKERS
-) as executor:
+with concurrent.futures.ProcessPoolExecutor(max_workers=DATASET_CREATION_NUM_WORKERS) as executor:
     while pending_samples != 0:
-
+        
         # reading in compressed samples to the buffer
         local_buffer = []
         local_buffer_size = min(pending_samples, DATASET_CREATION_BUFFER_SIZE)
         for i in range(local_buffer_size):
             local_buffer.append(all_samples.readline())
-
+            
         # process the datapoints stored in the buffer
         for processed_sample in executor.map(process_sample, local_buffer):
-
+            
             game_id, processed_datapoints = processed_sample
             # each game sample produces results for each phase and power
-            for (
-                is_validation_datapoint,
-                power_name,
-                game_result,
-            ) in processed_datapoints:
+            for is_validation_datapoint, power_name, game_result in processed_datapoints:
                 if game_result is not None:
-
-                    if is_validation_datapoint:
+                    
+                    if(is_validation_datapoint):
                         valid_dataset.write(compress_dict(game_result) + "\n")
-                        dataset_index["size_valid_dataset"] += 1
+                        dataset_index['size_valid_dataset'] += 1
                     else:
                         train_dataset.write(compress_dict(game_result) + "\n")
-                        dataset_index["size_train_dataset"] += 1
-
+                        dataset_index['size_train_dataset'] += 1
+                    
             progress_bar.update(phase_count_dataset.get(game_id))
             pending_samples -= 1
-
+            
         # free the memory of the buffer
         del local_buffer
+    
+# writing train/val datapoint counts 
+pickle.dump(dataset_index, open(MODEL_DATA_PATHS["TRAIN_VAL_DATA_COUNTS"], "wb"))           
 
-# writing train/val datapoint counts
-pickle.dump(dataset_index, open(MODEL_DATA_PATHS["TRAIN_VAL_DATA_COUNTS"], "wb"))
-
-progress_bar.close()
+progress_bar.close()    
 all_samples.close()
 train_dataset.close()
 valid_dataset.close()
